changed primitives

+++ app/src/main/java/com/mmdev/me/driver/data/datasource/local/fuel/history/entities/FuelHistoryEntity.kt
@@ -1,7 +1,7 @@
 /*
  * Created by Andrii Kovalchuk
  * Copyright (c) 2020. All rights reserved.
- * Last modified 17.08.2020 20:35
+ * Last modified 20.08.2020 18:34
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -24,7 +24,7 @@ import com.mmdev.me.driver.data.datasource.local.fuel.prices.entities.FuelStatio
 @Entity(tableName = "fuel_history")
 data class FuelHistoryEntity(
 	val distancePassed: Double,
-	val odometerValue: Double,
+	val odometerValue: Int,
 	@Embedded
 	val fuelStation: FuelStationEntity,
 	@Embedded
