+++ app/src/main/java/com/mmdev/me/driver/domain/fuel/history/model/FuelHistoryRecord.kt
@@ -1,7 +1,7 @@
 /*
  * Created by Andrii Kovalchuk
  * Copyright (c) 2020. All rights reserved.
- * Last modified 17.08.2020 20:35
+ * Last modified 20.08.2020 18:29
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -24,7 +24,7 @@ import java.util.*
 
 data class FuelHistoryRecord (
 	val distancePassed: Double = 0.0,
-	val odometerValue: Double = 0.0,
+	val odometerValue: Int = 0,
 	val fuelStation: FuelStation = FuelStation(),
 	val fuelPrice: FuelPrice = FuelPrice(),
 	val moneyCosts: Double = (fuelPrice.price * distancePassed).roundTo(2),
